word_cloud_app/requirements.txt
matplotlib==3.10.8
PyPDF2==3.0.1
python_docx==1.2.0
streamlit==1.53.1
wordcloud==1.9.6
